requirements.txt
python-telegram-bot==21.9
gspread==5.12.0
oauth2client==4.1.3
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
